+++ src/main/kotlin/com/duda/backbaseassignment/domain/model/Rating.kt
@@ -2,6 +2,6 @@ package com.duda.backbaseassignment.domain.model
 
 data class Rating(
     val rate: Short,
-    val user: Long,
+    val user: Int,
     val movie: String
 )
