+++ src/main/kotlin/com/duda/backbaseassignment/domain/service/exception/UserAlreadyRatedMovieException.kt
@@ -3,5 +3,5 @@ package com.duda.backbaseassignment.domain.service.exception
 import org.springframework.http.HttpStatus
 import org.springframework.web.server.ResponseStatusException
 
-class UserAlreadyRatedMovieException(movieTitle: String, userId: Long) :
+class UserAlreadyRatedMovieException(movieTitle: String, userId: Int) :
     ResponseStatusException(HttpStatus.BAD_REQUEST, "Movie $movieTitle was already rated by user $userId")
