+++ src/main/kotlin/com/duda/backbaseassignment/infra/database/RatingRepositoryAdapter.kt
@@ -2,8 +2,8 @@ package com.duda.backbaseassignment.infra.database
 
 import com.duda.backbaseassignment.domain.model.Rating
 import com.duda.backbaseassignment.domain.port.RatingRepository
-import com.duda.backbaseassignment.infra.database.records.Tables.MOVIE
-import com.duda.backbaseassignment.infra.database.records.Tables.RATING
+import com.duda.backbaseassignment.generated.Tables.MOVIE
+import com.duda.backbaseassignment.generated.Tables.RATING
 import org.jooq.DSLContext
 import org.springframework.stereotype.Repository
 
@@ -18,7 +18,7 @@ class RatingRepositoryAdapter(private val dslContext: DSLContext): RatingReposit
             .execute()
     }
 
-    override fun exists(movieTitle: String, userId: Long): Boolean {
+    override fun exists(movieTitle: String, userId: Int): Boolean {
         return dslContext.fetchExists(dslContext.selectOne()
             .from(RATING)
             .join(MOVIE)
